Add files via upload
I think this program would probably work. But, I am not sure because I am unable to test it out. Athar, you have the implementation for the sudoku puzzle. I have only assumed the name of an array containing the sudoku puzzle. It's getting late, would put any comments in it later.